hax new section header (WIP)
Some stuff are missing. Almost wrote the new section header (need to
compute some addresses and align them/adjust sizes).
After that it's almost done.
However it's a PoC, not a reliable virus, it's obvious to detect too.
This may be the base of a conception for a virus framework on
win32/win64. Hax everywhere to test stuff and probe PE.