Hax new section OK
New section but not valid data inside yet except for the name and maybe
the size of the data.

Before it failed because of the .exe~ file left by vim, I modified the
header to check if I could detect it. The program will fail if a .exe(~)
doesn't have a valid PE and DOS header.

Now when executed, it will add a section ".hax" to each .exe, it's
recommended to generate the virus each time with makeit.bat. It will
provide the DBG in ollydbg :D.

To test, copy ollydbg in the TP0 dir, it will get "infected" and you can
analyze it with PEView.

Haxhaxhax, almost done haxing.