OK hax
Works with gvim/peview/ollydbg but not with firefox though, I don't have
the time to debug and fix that. Maybe it's just that there is no room
for the new section header, dunno.

Infects all files in the directory. Each .exe will now infect again all
the files in its directory (and they'll do the same), and then the
original code is executed !
We don't check if a file is already infected, we would need to check
that to make it reliable.

Still, a virus !